updated form to be in a component
